requirements.txt
Django==2.0.4
django-archive==0.1.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.4
